pyproject.toml
[project]
name = "pea-momentum"
version = "0.1.0"
description = "Accelerated dual momentum on PEA-eligible UCITS ETFs"
authors = [{ name = "yohplala" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
license = { text = "MIT" }
dependencies = [
    "polars (>=1.20.0)",
    "yfinance (>=0.2.40)",
    "httpx (>=0.27.0)",
    "plotly (>=5.20.0)",
    "jinja2 (>=3.1.0)",
    "click (>=8.1.0)",
    "pyyaml (>=6.0.0)",
]

[project.scripts]
pea-mom = "pea_momentum.cli:cli"

[tool.poetry]
packages = [{ include = "pea_momentum" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.7.0"
mypy = "^1.13"
pre-commit = "^4.0.0"
types-pyyaml = "^6.0.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "B",        # bugbear
    "UP",       # pyupgrade
    "N",        # pep8-naming
    "RUF",      # ruff
    "SIM",      # simplify
    "PTH",      # pathlib
]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["yfinance.*", "plotly.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"
